+++ typescript/Custom.ts
@@ -56,7 +56,7 @@ function customStartSimulation() {
     totalCastSize = currentCast.length;
 
     let select = (<HTMLSelectElement>document.getElementById("format"));
-    //let select2 = (<HTMLSelectElement>document.getElementById("premiere-format"));
+    let select2 = (<HTMLSelectElement>document.getElementById("premiere-format"));
 
         if (select.options[select.selectedIndex].value == "top3")
             top3 = true;
@@ -72,18 +72,60 @@ function customStartSimulation() {
             allQueens = allQueens.filter(function (queen) {return currentCast.indexOf(queen) == -1});
         }
 
-        /*if (select2.options[select2.selectedIndex].value == "s6-premiere")
+        if (select2.options[select2.selectedIndex].value == "s6-premiere")
             s6Premiere = true;
         else if (select2.options[select2.selectedIndex].value == "s12-premiere")
-            s12Premiere = true;*/
+            s12Premiere = true;
+        else if (select2.options[select2.selectedIndex].value == "porkchop")
+            porkchopPremiere = true;
 
         if ((<HTMLInputElement>document.getElementById("disableDouble")).checked == true)
             noDouble = true;
         if ((<HTMLInputElement>document.getElementById("disableReturn")).checked == true)
             noReturn = true;
 
-        if (currentCast.length == 3 && top4 || currentCast.length == 3 && all_stars)
+        if (currentCast.length == 3 && top4 || currentCast.length == 3 && all_stars) 
             window.alert("Lip-Sync For The Crown and All Star formats needs at least 4 queens!");
+        else if ((s12Premiere || s6Premiere || porkchopPremiere) && currentCast.length < 10) 
+            window.alert("You need at least 10 queens for a double premiere!");        
+        else if (s12Premiere || s6Premiere)
+            doublePremiere();
+        else if (porkchopPremiere)
+            porkchopLipsyncs();
         else
             newEpisode();
+}
+
+function preQueens(): void {
+    let select = <HTMLSelectElement>document.getElementById("preQueens");
+
+    for (let i = 0; i < allQueens.length; i++)
+    {
+        let option = document.createElement("option");
+        option.innerHTML = allQueens[i].getName();
+        select.appendChild(option);
+    }
+}
+
+function addPreQueen(): void {
+    let select = <HTMLSelectElement>document.getElementById("preQueens");
+    let value = select.options[select.selectedIndex].value;
+
+    for (let i = 0; i < allQueens.length; i++) {
+        if (value == allQueens[i].getName()) {
+            customCast.push(allQueens[i]);
+
+            let list = document.getElementById("cast");
+            let queen = document.createElement("option");
+            queen.text = allQueens[i].getName();
+            list!.appendChild(queen);
+        }
+    }
+}
+
+function randomizeStats(): void {
+    let stats = <HTMLCollection>document.getElementsByClassName("stats");
+    for (let i = 0; i < stats.length; i++) {
+        (<HTMLInputElement>stats[i]).value = randomNumber(0, 15).toString();
+    }
 }