+++ typescript/DoublePremiere.ts
@@ -1,0 +1,125 @@
+let premiereCounter: number = 0;
+
+let firstCast: Array<Queen> = [];
+let secondCast: Array<Queen> = [];
+
+function doublePremiere() {
+    if (premiereCounter == 0)
+        if (s6Premiere || s12Premiere) {
+            shuffle(currentCast)
+
+            firstCast = currentCast.splice(0, Math.floor(currentCast.length / 2));
+            secondCast = [...currentCast];
+        }
+    
+    if (premiereCounter == 0) {
+        currentCast = firstCast;
+        for (let i = 0; i < secondCast.length; i++)
+            secondCast[i].addToTrackRecord("");
+        premiereCounter++;
+        newEpisode();
+    } else if (premiereCounter == 1) {
+        currentCast = secondCast;
+        for (let i = 0; i < firstCast.length; i++)
+            firstCast[i].addToTrackRecord("");
+        premiereCounter++
+        newEpisode();
+    } else if (premiereCounter == 2) {
+        currentCast = [...firstCast, ...secondCast];
+        premiereCounter++;
+        newEpisode();
+    }
+}
+
+function porkchopLipsyncs(): void {
+    let screen = new Scene();
+    screen.clean();
+    screen.createHeader("It's time...");
+    screen.createParagraph("After the queens enter the workroom, it's time for them to lip-sync... for their lives!");
+    
+    for (let i = 0; i < Math.floor(totalCastSize / 2); i++) {
+        screen.createHorizontalLine();
+
+        let queen1 = currentCast[randomNumber(0, currentCast.length - 1)];
+        currentCast.splice(currentCast.indexOf(queen1), 1);
+        let queen2 = currentCast[randomNumber(0, currentCast.length - 1)];
+        currentCast.splice(currentCast.indexOf(queen2), 1);
+        if (currentCast.length == 1) {
+            let queen3 = currentCast[randomNumber(0, currentCast.length - 1)];
+            currentCast.splice(currentCast.indexOf(queen3), 1);
+        
+            screen.createBold(`${queen1.getName()}, ${queen2.getName()} and ${queen3.getName()} will lipsync...`);
+            lsSong();
+            let lipSync = [queen1, queen2, queen3];
+
+            for (let i = 0; i < lipSync.length; i++) {
+                lipSync[i].getASLipsync();
+            }
+            lipSync.sort((a, b) => (b.lipsyncScore - a.lipsyncScore));
+    
+            screen.createBold(`${queen1.getName()}, shantay you stay!`);
+            screen.createBold(`${queen2.getName()} and ${queen3.getName()}, you're getting the porkchop...`);
+
+            queen1.addToTrackRecord(" WIN ");
+            queen2.addToTrackRecord("LOSS");
+            queen3.addToTrackRecord("LOSS")
+
+            firstCast.push(queen1);
+            secondCast.push(queen2, queen3);
+        } else { 
+            screen.createBold(`${queen1.getName()} and ${queen2.getName()} will lipsync...`);
+            lsSong();
+
+            let lipSync = [queen1, queen2];
+
+            for (let i = 0; i < lipSync.length; i++) {
+                lipSync[i].getASLipsync();
+            }
+            lipSync.sort((a, b) => (b.lipsyncScore - a.lipsyncScore));
+
+            screen.createBold(`${queen1.getName()}, shantay you stay!`);
+            screen.createBold(`${queen2.getName()}, you're getting the porkchop...`);
+
+            queen1.addToTrackRecord(" WIN ");
+            queen2.addToTrackRecord("LOSS");
+
+            firstCast.push(queen1);
+            secondCast.push(queen2);
+        }
+    }
+
+    episodeChallenges.push("Porkchop");
+
+    screen.createButton("Proceed", "doublePremiere()");
+}
+
+function doublePremiereJudging() {
+    let screen = new Scene();
+
+	screen.clean();
+	screen.createHeader("Bring back my girls!");
+
+	screen.createBold("Ladies, I've made some decisions...");
+
+    screen.createBold(`${topQueens[0].getName()}, ${topQueens[1].getName()}, condragulations, you're the Top 2 of the week!`);
+    screen.createParagraph("Nobody is going home tonight!");
+    screen.createHorizontalLine();
+
+    screen.createBold("The Top 2 will now lip-sync... for the win!");
+    lsSong();
+
+    for (let i = 0; i < topQueens.length; i++) {
+        topQueens[i].getASLipsync();
+    }
+    topQueens.sort((a, b) => (b.lipsyncScore - a.lipsyncScore));
+
+    screen.createBold(`${topQueens[0].getName()}, you're a winner baby!`);
+
+    topQueens[0].addToTrackRecord("WIN");
+    topQueens[0].favoritism += 5;
+
+    topQueens[1].addToTrackRecord("TOP2");
+    topQueens[0].favoritism += 2;
+
+    screen.createButton("Proceed", "doublePremiere()");
+}